remove search container and replace connect to search component, delete container folder
