--- src/containers/Search.js
@@ -1,15 +1,0 @@
-import { connect } from 'react-redux';
-import Search from '../components/Search';
-
-const mapStateToProps = ({ search }) => ({
-  ...search,
-});
-
-const mapDispatchToProps = dispatch => ({
-  getActionDispatcher: action => () => dispatch(action),
-});
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(Search);
